+++ src/stuart/WebHdfsFileSystem.lua
@@ -13,23 +13,51 @@ function WebHdfsFileSystem:initialize(uri)
   self.parsedUri = url.parse(uri)
 end
 
+function WebHdfsFileSystem:getFileStatus(path)
+  local urlSegments = moses.clone(self.parsedUri)
+  urlSegments.path = urlSegments.path .. '/v1/' .. (path or '')
+  urlSegments.query = 'op=GETFILESTATUS'
+  local uri = url.build(urlSegments)
+  local json, status, headers = http.request(uri)
+  local obj = jsonutil.decode(json)
+  if obj.RemoteException then error(obj.RemoteException.message) end
+  return obj.FileStatus, status, headers
+end
+
+function WebHdfsFileSystem:isDirectory(path)
+  local fileStatus = self:getFileStatus(path)
+  return fileStatus and fileStatus.type == 'DIRECTORY'
+end
+
 function WebHdfsFileSystem:listStatus(path)
   local urlSegments = moses.clone(self.parsedUri)
-  urlSegments.path = urlSegments.path .. path
+  urlSegments.path = urlSegments.path .. '/v1/' .. (path or '')
   urlSegments.query = 'op=LISTSTATUS'
   local uri = url.build(urlSegments)
-  local json = http.request(uri)
+  local json, status, headers = http.request(uri)
   local obj = jsonutil.decode(json)
-  return obj.FileStatuses.FileStatus
+  if obj.RemoteException then error(obj.RemoteException.message) end
+  return obj.FileStatuses.FileStatus, status, headers
+end
+
+function WebHdfsFileSystem:mkdirs(path)
+  local urlSegments = moses.clone(self.parsedUri)
+  urlSegments.path = urlSegments.path .. '/v1/' .. (path or '')
+  urlSegments.query = 'op=MKDIRS'
+  local uri = url.build(urlSegments)
+  local json, status, headers = http.request(uri)
+  local obj = jsonutil.decode(json)
+  if obj.RemoteException then error(obj.RemoteException.message) end
+  return obj.boolean, status, headers
 end
 
 function WebHdfsFileSystem:open(path)
   local urlSegments = moses.clone(self.parsedUri)
-  urlSegments.path = urlSegments.path .. path
+  urlSegments.path = urlSegments.path .. '/v1/' .. (path or '')
   urlSegments.query = 'op=OPEN'
   local uri = url.build(urlSegments)
-  local data = http.request(uri)
-  return data
+  local data, status, headers = http.request(uri)
+  return data, status, headers
 end
 
 return WebHdfsFileSystem